+++ packages/uikit-prerenderer/index.js
@@ -1,6 +1,6 @@
 const promiseLimit = require('promise-limit');
 
-const puppeteer = require('puppeteer-core');
+const puppeteer = require('puppeteer');
 
 const localChrome = require('local-chrome');